refactor: (#20) 피드백 반영
- 변수명, 파일명 명확하게 변경
- 유효성 검사 미들웨어(validationMiddleware) 추가
- toCamelCase 변환 로직을 레포지토리로 이동
- HTTP 데이터 처리를 컨트롤러에서 수행하도록 변경
- 댓글 좋아요(reaction) API 메서드를 PUT으로 변경
- params 유효성 검사 추가

+++ app/src/routes/comment/commentController.js
@@ -5,8 +5,17 @@ const CommentService = require("../../services/comment/commentService.js");
 module.exports = {
   createComment: async (req, res) => {
     try {
+      const { episodeId } = req.params;
+      const { content, parentId } = req.body;
+      const userId = req.user.id;
+
       const commentService = new CommentService(req);
-      const { status, success, data } = await commentService.createComment();
+      const { status, success, data } = await commentService.createComment(
+        episodeId,
+        content,
+        parentId,
+        userId
+      );
 
       return res.status(status).json({ success, data });
     } catch (error) {
@@ -20,8 +29,16 @@ module.exports = {
 
   updateComment: async (req, res) => {
     try {
+      const { commentId } = req.params;
+      const { content } = req.body;
+      const userId = req.user.id;
+
       const commentService = new CommentService(req);
-      const { status, success, data } = await commentService.updateComment();
+      const { status, success, data } = await commentService.updateComment(
+        commentId,
+        content,
+        userId
+      );
 
       return res.status(status).json({ success, data });
     } catch (error) {
@@ -35,12 +52,15 @@ module.exports = {
 
   deleteComment: async (req, res) => {
     try {
+      const { commentId } = req.params;
+      const userId = req.user.id;
+
       const commentService = new CommentService(req);
-      const { status, success, data } = await commentService.deleteComment();
+      const { status, success, data } = await commentService.deleteComment(commentId, userId);
 
       return res.status(status).json({ success, data });
     } catch (error) {
-      console.log("deleteComment error: ", error);
+      console.error("deleteComment error: ", error);
       return res.status(500).json({
         success: false,
         data: { message: "서버 오류가 발생했습니다." },
@@ -50,8 +70,12 @@ module.exports = {
 
   reactComment: async (req, res) => {
     try {
+      const { commentId } = req.params;
+      const { type } = req.body;
+      const userId = req.user.id;
+
       const commentService = new CommentService(req);
-      const { status, success, data } = await commentService.reactComment();
+      const { status, success, data } = await commentService.reactComment(commentId, type, userId);
 
       return res.status(status).json({ success, data });
     } catch (error) {
